requirements.txt
tensorflow-gpu
tb-nightly
future
pandas
matplotlib
nltk
colorama==0.4.1
decorator==4.3.2
entrypoints==0.3
ipykernel==5.1.0
ipython==7.3.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
jedi==0.13.3
Jinja2==2.10
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.4.0
MarkupSafe==1.1.1
mistune==0.8.4
mkl-fft==1.0.10
mkl-random==1.0.2
nbconvert==5.3.1
nbformat==4.4.0
notebook==5.7.4
numpy==1.16.2
olefile==0.46
pandocfilters==1.4.2
parso==0.3.4
pickleshare==0.7.5
Pillow==5.4.1
prometheus-client==0.6.0
prompt-toolkit==2.0.9
pycparser==2.19
Pygments==2.3.1
python-dateutil==2.8.0
pywinpty==0.5.5
PyYAML==5.1
pyzmq==18.0.0
qtconsole==4.4.3
scikit-learn==0.20.2
scipy==1.2.1
Send2Trash==1.5.0
sip==4.19.13
six==1.12.0
terminado==0.8.1
testpath==0.4.2
torch==1.0.1
torchvision==0.2.2
tornado==5.1.1
tqdm==4.31.1
traitlets==4.3.2
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==3.4.2
wincertstore==0.2
